+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * AudioPlayer bug fixed
 * Placeholders updated
 * New Artworks added
+* `CachedPhoto` bug fixed
 * Bug fixes and improvements
 
 ## 0.29.5